Fix to handle float values and then downcast to int.

+++ server/ParkingIQServerApp/src/main/java/edu/colorado/piq/service/impl/ParkingLotReportServiceImpl.java
@@ -54,7 +54,7 @@ public class ParkingLotReportServiceImpl implements ParkingLotReportService {
 				String[] keys = column.getName().split("_");
 				Day day = Day.parseDay(keys[0]);
 				Session session = Session.parseSession(keys[1]);
-				int availableSpace = Integer.parseInt(column.getValue());
+				int availableSpace = (int) Float.parseFloat(column.getValue());
 				report.addDayReport(day, session, availableSpace);
 			}
 		}
